# cmd/docs/swagger.yaml
basePath: /api/v1
definitions:
  domain.AddSubcriptionRequest:
    properties:
      end_date:
        example: 07-2026
        type: string
      price:
        example: 400
        type: integer
      service_name:
        example: Yandex Plus
        type: string
      start_date:
        example: 07-2025
        type: string
      user_id:
        example: a19df875-4040-4fc3-84ad-003d013fcd89
        type: string
    required:
    - price
    - service_name
    - start_date
    - user_id
    type: object
  domain.Subscription:
    properties:
      endDate:
        example: 07-2026
        type: string
      id:
        type: string
      price:
        example: 400
        type: integer
      serviceName:
        example: Yandex Plus
        type: string
      startDate:
        example: 07-2025
        type: string
      userID:
        example: a19df875-4040-4fc3-84ad-003d013fcd89
        type: string
    type: object
  domain.UpdateSubcriptionRequest:
    properties:
      end_date:
        example: 07-2026
        type: string
      id:
        type: string
      price:
        example: 400
        type: integer
      service_name:
        example: Yandex Plus
        type: string
      start_date:
        example: 07-2025
        type: string
      user_id:
        example: a19df875-4040-4fc3-84ad-003d013fcd89
        type: string
    required:
    - id
    - price
    - service_name
    - start_date
    - user_id
    type: object
host: localhost:8080
info:
  contact: {}
  description: API для управления подписками
  title: Subscribe API
  version: "1.0"
paths:
  /{id}:
    delete:
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
      summary: Удалить подписку
    get:
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Subscription'
      summary: Получить подписку
  /all:
    get:
      parameters:
      - description: Описание параметра
        in: query
        name: param_name
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Получить все подписки
  /create:
    post:
      parameters:
      - description: Данные подписки
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/domain.AddSubcriptionRequest'
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Создать подписку
  /total:
    get:
      parameters:
      - description: ID пользователя
        in: query
        name: user_id
        type: string
      - description: Название сервиса
        in: query
        name: service_name
        type: string
      - description: Начальная дата (YYYY-MM-DD)
        in: query
        name: start_date
        required: true
        type: string
      - description: Конечная дата (YYYY-MM-DD)
        in: query
        name: end_date
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Подсчет суммарной стоимости всех подписок за выбранный период с фильтрацией
        по id пользователя и названию подписки
  /update:
    put:
      parameters:
      - description: Данные
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateSubcriptionRequest'
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Изменить подписку
swagger: "2.0"